Updated all games to use floating point 0-1 when dealing with sector light levels. Part #3

+++ doomsday/plugins/wolftc/include/r_defs.h
@@ -76,7 +76,7 @@ typedef struct xsector_s {
         float       origheight;
     } planes[2];    // {floor, ceiling}
 
-    int             origlight;
+    float           origlight;
     byte            origrgb[3];
     xgsector_t     *xg;
